requirements.txt
numpy
Pillow
tensorflow==1.14.0
unittest
